+++ CHANGELOG.md
@@ -6,10 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `@distributeedlab/w3p` - EVM fallback provider, readme details about custom provider implementation
+- `@distributeedlab/w3p` - EVM fallback provider, readme details about custom provider implementation, possibility to set details for many supported chains
 
 ### Changed
-- `@distributeedlab/w3p` - `multiple-providers` examples about adding details for supported chains
+- `@distributeedlab/w3p` - `multiple-providers` examples about adding details for supported chains, `SwitchChain` method of `ProviderBase` interface now optional
 
 ## [0.2.0-rc.10] - 2023-05-11
 ### Changed